dbus: refactor each interface into its own file
