+++ src/comm/dbus/mod.rs
@@ -10,4 +10,6 @@
 //
 // You should have received a copy of the GNU General Public License along with this program.  If not, see http://www.gnu.org/licenses/.
 
-pub mod interfaces;
+pub mod status_interface;
+pub mod control_interface;
+pub mod configure_interface;
